feat: add isolated apis as exported
