+++ src/index.ts
@@ -1,1 +1,29 @@
-export { TMDB } from './tmdb/TMDB'
+// Isolated APIs
+export { Account } from './tmdb/apis/account'
+export { Authentication } from './tmdb/apis/authentication'
+export { Certifications } from './tmdb/apis/certifications'
+export { Changes } from './tmdb/apis/changes'
+export { Collections } from './tmdb/apis/collections'
+export { Companies } from './tmdb/apis/companies'
+export { Configuration } from './tmdb/apis/configuration'
+export { Credits } from './tmdb/apis/credits'
+export { Discover } from './tmdb/apis/discover'
+export { Find } from './tmdb/apis/find'
+export { Genres } from './tmdb/apis/genres'
+export { GuestSessions } from './tmdb/apis/guest-sessions'
+export { Keywords } from './tmdb/apis/keywords'
+export { Lists } from './tmdb/apis/lists'
+export { Movies } from './tmdb/apis/movies'
+export { Networks } from './tmdb/apis/networks'
+export { People } from './tmdb/apis/people'
+export { Reviews } from './tmdb/apis/reviews'
+export { Search } from './tmdb/apis/search'
+export { Trending } from './tmdb/apis/trending'
+export { TV } from './tmdb/apis/tv'
+export { TVEpisodes } from './tmdb/apis/tv-episodes'
+export { TVSeasons } from './tmdb/apis/tv-seasons'
+export { TVEpisodeGroups } from './tmdb/apis/tv-episode-groups'
+export { WatchProviders } from './tmdb/apis/watch-providers'
+
+// Wrapper
+export { Wrapper as TMDB } from './tmdb/wrapper'
