+++ src/tmdb/apis/collections/index.ts
@@ -1,4 +1,4 @@
-import { BaseAPI } from '../BaseAPI'
+import { BaseAPI } from '../../base-api'
 import {
   CollectionsFilters,
   CollectionDetailsResponse,
