+++ src/tmdb/apis/companies/index.ts
@@ -1,4 +1,4 @@
-import { BaseAPI } from '../BaseAPI'
+import { BaseAPI } from '../../base-api'
 import {
   DetailsResponse,
   ImagesResponse,
