+++ src/tmdb/apis/genres/index.ts
@@ -1,4 +1,4 @@
-import { BaseAPI } from '../BaseAPI'
+import { BaseAPI } from '../../base-api'
 import { GenresFilters, GenresResponse } from './types'
 
 export class Genres extends BaseAPI {
