+++ src/tmdb/apis/guest-sessions/index.ts
@@ -1,4 +1,4 @@
-import { BaseAPI } from '../BaseAPI'
+import { BaseAPI } from '../../base-api'
 import {
   RatedFilters,
   RatedMoviesResponse,
