+++ src/tmdb/apis/lists/index.ts
@@ -1,4 +1,4 @@
-import { BaseAPI } from '../BaseAPI'
+import { BaseAPI } from '../../base-api'
 import {
   AddItemBody,
   AddItemFilters,
