+++ src/tmdb/wrapper/index.ts
@@ -1,33 +1,31 @@
-import { TMDBOptions } from '../types'
-import { Certifications } from './apis/certifications'
-import { Changes } from './apis/changes'
-import { Collections } from './apis/collections'
-import { Companies } from './apis/companies'
-import { Configuration } from './apis/configuration'
-import { Credits } from './apis/credits'
-import { Discover } from './apis/discover'
-import { Genres } from './apis/genres'
-import { Movies } from './apis/movies'
-import { Trending } from './apis/trending'
-import { Find } from './apis/find'
-import { GuestSessions } from './apis/guest-sessions'
-import { Keywords } from './apis/keywords'
-import { Account } from './apis/account'
-import { Authentication } from './apis/authentication'
-import { Lists } from './apis/lists'
-import { Networks } from './apis/networks'
-import { People } from './apis/people'
-import { Reviews } from './apis/reviews'
-import { Search } from './apis/search'
-import { TV } from './apis/tv'
-import { WatchProviders } from './apis/watch-providers'
-import { TVEpisodes } from './apis/tv-episodes'
-import { TVSeasons } from './apis/tv-seasons'
-import { TVEpisodeGroups } from './apis/tv-episode-groups'
+import { TMDBOptions } from '../../types'
+import { Certifications } from '../apis/certifications'
+import { Changes } from '../apis/changes'
+import { Collections } from '../apis/collections'
+import { Companies } from '../apis/companies'
+import { Configuration } from '../apis/configuration'
+import { Credits } from '../apis/credits'
+import { Discover } from '../apis/discover'
+import { Genres } from '../apis/genres'
+import { Movies } from '../apis/movies'
+import { Trending } from '../apis/trending'
+import { Find } from '../apis/find'
+import { GuestSessions } from '../apis/guest-sessions'
+import { Keywords } from '../apis/keywords'
+import { Account } from '../apis/account'
+import { Authentication } from '../apis/authentication'
+import { Lists } from '../apis/lists'
+import { Networks } from '../apis/networks'
+import { People } from '../apis/people'
+import { Reviews } from '../apis/reviews'
+import { Search } from '../apis/search'
+import { TV } from '../apis/tv'
+import { WatchProviders } from '../apis/watch-providers'
+import { TVEpisodes } from '../apis/tv-episodes'
+import { TVSeasons } from '../apis/tv-seasons'
+import { TVEpisodeGroups } from '../apis/tv-episode-groups'
 
-export class TMDB {
-  static API_URL_V3 = 'https://api.themoviedb.org/3'
-
+export class Wrapper {
   public account: Account
   public authentication: Authentication
   public certifications: Certifications
@@ -55,10 +53,6 @@ export class TMDB {
   public watchProviders: WatchProviders
 
   constructor(options: TMDBOptions) {
-    if (!options.url) {
-      options.url = TMDB.API_URL_V3
-    }
-
     this.account = new Account(options)
     this.authentication = new Authentication(options)
     this.certifications = new Certifications(options)
